# server/docs/swagger.yaml
info:
  contact: {}
paths:
...
  /admin/complete-table:
    get:
      description: '"结台"'
      parameters:
      - description: table_id
        in: query
        name: table_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: json
      summary: CompleteTable
      tags:
      - public
  /admin/open-table:
    get:
      description: '"开台"'
      parameters:
      - description: table_id
        in: query
        name: table_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: json
      summary: OpenTable
      tags:
      - public
...
swagger: "2.0"